Removed old Login screen and replaced with Jerome's
