+++ client/src/views/Login.js
@@ -1,39 +1,104 @@
-import * as React from "react";
-import "../css/Common.css";
-import Card from "../components/Card";
-import Grid from "@mui/material/Grid";
-import MenuIcon from "@mui/icons-material/Menu";
-import Box from "@mui/material/Box";
-import { FormControl } from "@mui/material";
-import Select from "@mui/material/Select";
-import TextField from "@mui/material/TextField";
+import * as React from 'react';
+import Avatar from '@mui/material/Avatar';
+import Button from '@mui/material/Button';
+import CssBaseline from '@mui/material/CssBaseline';
+import TextField from '@mui/material/TextField';
+import FormControlLabel from '@mui/material/FormControlLabel';
+import Checkbox from '@mui/material/Checkbox';
+import Link from '@mui/material/Link';
+import Paper from '@mui/material/Paper';
+import Box from '@mui/material/Box';
+import Grid from '@mui/material/Grid';
+import Typography from '@mui/material/Typography';
+import { createTheme, ThemeProvider } from '@mui/material/styles';  
+
+const defaultTheme = createTheme();
 
 const Login = () => {
-  return (
-    <div>
-      <div className="gradient_background">
-        <Grid container columns={2} sx={{ p: 5 }}>
-          <Grid item xs={1}>
-            <h>Hamlet.</h>
-          </Grid>
-
-          <Grid item xs={1} sx={{ p: 5 }}>
-            <Box display="flex" justifyContent="flex-end">
-              <MenuIcon style={{ color: "white" }} fontSize="large" />
+    return (
+      <ThemeProvider theme={defaultTheme}>
+      <Grid container component="main" sx={{ height: '100vh' }}>
+        <CssBaseline />
+        <Grid
+          item
+          xs={false}
+          sm={4}
+          md={7}
+          sx={{
+            backgroundImage: 'url(https://source.unsplash.com/random?wallpapers)',
+            backgroundRepeat: 'no-repeat',
+            backgroundColor: (t) =>
+              t.palette.mode === 'light' ? t.palette.grey[50] : t.palette.grey[900],
+            backgroundSize: 'cover',
+            backgroundPosition: 'center',
+          }}
+        />
+        <Grid item xs={12} sm={8} md={5} component={Paper} elevation={6} square>
+          <Box
+            sx={{
+              my: 8,
+              mx: 4,
+              display: 'flex',
+              flexDirection: 'column',
+              alignItems: 'center',
+            }}
+          >
+            <Avatar sx={{ m: 1, bgcolor: 'secondary.main' }}>
+            </Avatar>
+            <Typography component="h1" variant="h5">
+              Sign in
+            </Typography>
+            <Box component="form" sx={{ mt: 1 }}>
+              <TextField
+                margin="normal"
+                required
+                fullWidth
+                id="email"
+                label="Email Address"
+                name="email"
+                autoComplete="email"
+                autoFocus
+              />
+              <TextField
+                margin="normal"
+                required
+                fullWidth
+                name="password"
+                label="Password"
+                type="password"
+                id="password"
+                autoComplete="current-password"
+              />
+              <FormControlLabel
+                control={<Checkbox value="remember" color="primary" />}
+                label="Remember me"
+              />
+              <Button
+                type="submit"
+                fullWidth
+                variant="contained"
+                sx={{ mt: 3, mb: 2 }}
+              >
+                Sign In
+              </Button>
+              <Grid container>
+                <Grid item xs>
+                  <Link href="#" variant="body2">
+                    Forgot password?
+                  </Link>
+                </Grid>
+                <Grid item>
+                  <Link href="#" variant="body2">
+                    {"Don't have an account? Sign Up"}
+                  </Link>
+                </Grid>
+              </Grid>
             </Box>
-          </Grid>
+          </Box>
         </Grid>
-          <Grid container>
-            <FormControl fullWidth sx={{ p: 4, pt: 10, minWidth: 120 }}>
-              <TextField required fullWidth label="Username" id="fullWidth" />
-            </FormControl>
-            <FormControl fullWidth sx={{ p: 4, pt: 0, minWidth: 120 }}>
-              <TextField required fullWidth label="Password" id="fullWidth" />
-            </FormControl>
-          </Grid>
-      </div>
-    </div>
-  );
-};
-
-export default Login;
+      </Grid>
+    </ThemeProvider>
+    ) 
+  };
+  
+  export default Login;
